Fix permissions and imports for actions
Also add documentation for video readyState and format prisma schema

+++ features/webcams/index.ts
@@ -3,18 +3,22 @@ import {
   editWebcamSchema,
   removeWebcamSchema,
 } from "@/app/(authenticated)/webcam/schema";
+import { requirePermission } from "@/lib/auth/server";
 import { prisma } from "@/lib/db";
 import { z } from "zod";
 
 export async function fetchWebcamFeeds() {
+  await requirePermission("Webcams.View");
   return await prisma.webcamFeed.findMany();
 }
 
 export async function addWebcamFeed(data: z.infer<typeof addWebcamSchema>) {
+  await requirePermission("Webcams.Manage");
   return await prisma.webcamFeed.create({ data });
 }
 
 export async function editWebcamFeed(data: z.infer<typeof editWebcamSchema>) {
+  await requirePermission("Webcams.Manage");
   return await prisma.webcamFeed.update({
     where: { webcam_id: data.webcam_id },
     data: {
@@ -28,6 +32,7 @@ export async function editWebcamFeed(data: z.infer<typeof editWebcamSchema>) {
 export async function removeWebcamFeed(
   data: z.infer<typeof removeWebcamSchema>,
 ) {
+  await requirePermission("Webcams.Manage");
   return await prisma.webcamFeed.delete({
     where: {
       webcam_id: data.webcam_id,
